Add logic to Library to find Memory Leaks
Memory leaks are a nasty problem.  VS17 gives some help but identifies the offending nodes by location in memory.  The compiler provides some left handed help in the form of file name and line number but the "new" and "delete" c++ don't provide these identifiers.  So a kludge was re-invented (probably) to record when debugging the file name and line number with each allocated item.  VS17 reports each memory leak when debugging with the source name/line number information and bob's you uncle!